version 3 - filtros

+++ Presentacion/FrmPrincipal.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+    
 
 namespace Presentacion
 {
@@ -24,7 +25,10 @@ namespace Presentacion
         private void FrmPrincipal_Load(object sender, EventArgs e)
         {
             cargarBase();
-
+            cargarListaCampo();
+            cargarListaCriterio(cbxCampo.Text);
+            
+            
             
   
         }
@@ -89,7 +93,15 @@ namespace Presentacion
 
             ArticuloNegocio articuloNegocio = new ArticuloNegocio();
 
-            articuloNegocio.eliminarArticulo(articuloSeleccionado);
+            DialogResult resultado = MessageBox.Show("Seguro desea eliminar el articulo seleccionado?", "Advertencia", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+
+            if (resultado == DialogResult.Yes)
+            {
+                articuloNegocio.eliminarArticulo(articuloSeleccionado);
+            }
+            
+
+            
 
 
             cargarBase();
@@ -101,6 +113,49 @@ namespace Presentacion
             frmAlta.ShowDialog();
             cargarBase();
 
+        }
+
+        private void gbxFiltro_Enter(object sender, EventArgs e)
+        {
+
+        }
+
+        private void cargarListaCampo() 
+        {
+            List<String> campos = new List<String>();
+            campos.Add("Codigo");
+            campos.Add("Nombre");
+            campos.Add("Descripcion");
+            campos.Add("Marca");
+            campos.Add("Categoria");
+            campos.Add("Precio");
+
+            cbxCampo.DataSource = campos;
+
+        }
+        private void cargarListaCriterio(String campo) 
+        {
+            List<String> listaCriterios = new List<String>();
+
+            if (campo == "Precio")
+            {
+                listaCriterios.Add("Mayor a");
+                listaCriterios.Add("Menor a");
+                listaCriterios.Add("Igual a");
+            }
+            else 
+            {
+                listaCriterios.Add("Contiene");
+                listaCriterios.Add("Empieza con");
+                listaCriterios.Add("Termina con");
+            }
+
+            cbxCriterio.DataSource = listaCriterios;
+        }
+
+        private void cbxCampo_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            cargarListaCriterio(cbxCampo.Text);
         }
     }
 }
